gate manageCategories para admin

+++ app/Http/Service/CategoryService.php
@@ -3,6 +3,7 @@
 namespace App\Http\Service;
 
 use App\Http\Repository\CategoryRepository;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryService {
     public function __construct(protected CategoryRepository $repository) {}
@@ -16,14 +17,20 @@ class CategoryService {
     }
 
     public function newCategory(array $data) {
+        Gate::authorize("manageCategories");
+
         return $this->repository->newCategory($data);
     }
 
     public function updateCategory(array $data, $id) {
+        Gate::authorize("manageCategories");
+
         return $this->repository->updateCategory($data, $id);
     }
 
     public function deleteCategory($id) {
+        Gate::authorize("manageCategories");
+
         return $this->repository->deleteCategory($id);
     }
 }
